+++ src/pages/components/homePage/mainSection/mainProductSection.scss
@@ -1,0 +1,16 @@
+@import '../../../mixin';
+
+div.content{
+    @include initial;
+    @include flexDisplay(center, normal);
+    aside.cat {
+        @include border();
+        @include initial;
+        width: 16rem;
+    }
+    div.product {
+        display: block;
+        width: 60rem;
+        float: left;
+    }
+}
