+++ src/pages/components/homePage/mainSection/molecules/itemPrice.scss
@@ -1,0 +1,54 @@
+@import '../../../../mixin';
+
+div.price {
+    @include flexDisplay(space-between);
+    padding: 0.8rem;
+    
+    div.value {
+        text-align: center;
+        p {
+            font-size: 1rem;
+            margin: 0;
+        }
+        s {
+            font-size: 0.8rem;
+            color: gray;
+        }
+    }
+}
+
+button.addBtn {
+    width: 5rem;
+    font-size: 1rem;
+    color: green;
+    background-color: rgb(214, 246, 231);
+    @include border(0.03125rem, green);
+    border-radius: $borderValue*10;
+    height: 2rem;
+    cursor: pointer;
+}
+div.billItemBtn {
+    @include flexDisplay();
+    margin-left: auto;
+    margin-right: 0;
+    width: 5rem;
+    div, button {
+        border: none;
+        color: white;
+        background-color: green;
+        width: 5rem;
+        height: 2rem;
+        font-size: 1.05rem;
+        @include flexDisplay();
+
+        &:nth-child(1) {
+            border-radius: 0.4rem 0 0 0.4rem;
+            cursor: pointer;
+        }
+
+        &:nth-child(3) {
+            border-radius: 0 0.4rem 0.4rem 0;
+            cursor: pointer;
+        }
+    }
+}
