+++ src/pages/components/homePage/mainSection/molecules/productHead.scss
@@ -1,0 +1,28 @@
+@import '../../../../mixin';
+
+
+header.productHead {
+    @include flexDisplay(space-between);
+    @include border();
+    h2 {
+        font-size: 0.9rem;
+        padding-left: 1rem;
+    }
+    div.filter {
+        display: flex;
+        width: 18rem;
+        align-items: center;
+        font-size: 0.9rem;
+        color: rgb(99, 99, 99);
+
+        select#filterBar {
+            width: 10.5rem;
+            padding: 0 0.3rem;
+            margin: 0.9rem;
+            height: 1.7rem;
+            color: green;
+            @include border;
+            border-radius: $borderValue*10;
+        }
+    }
+}
