+++ src/pages/components/homePage/mainSection/molecules/productItem.scss
@@ -1,0 +1,61 @@
+@import '../../../../mixin';
+
+section.container {
+    @include flexDisplay(left);
+    @include border();
+    flex-wrap: wrap;
+
+    div.item {
+        @include border();
+        width: 14.92rem;
+
+        button.discount {
+            color: white;
+            border: none;
+            font-size: 0.8rem;
+            padding: 0.3rem;
+            margin: 0.2rem;
+            border-radius: $borderValue*10;
+            background-color: rgb(78, 114, 233);
+            cursor: pointer;
+        }
+        
+        div.itemImgDiv {
+            @include flexDisplay();
+        
+            img.itemImg {
+                height: 8rem;
+                display: flex;
+                justify-content: center;
+            }
+        }
+
+        div.itemSourcedTime {
+            @include flexDisplay();
+            width: 7.5rem;
+            height: 1rem;
+            margin: 0.2rem 0.2rem 0.2rem auto;
+            padding: 0.2rem;
+            font-size: 0.8rem;
+            color: green;
+            background-color: rgb(214, 246, 231);
+            border-radius: $borderValue*10;
+            cursor: pointer;
+        
+            img.sourcedIcon {
+                padding: 0 0.4rem;
+                width: 0.7rem;
+            }
+        }
+        
+        div.itemName {
+            font-size: 1rem;
+            padding: 0.3rem 0.8rem;
+        }
+        div.itemWeight {
+            font-size: 0.9rem;
+            padding: 0.75rem;
+            color: gray;
+        }
+    } 
+}
