--- src/pages/components/homePage/molecules/services/ourServices.scss
@@ -1,40 +1,0 @@
-@import '../../../../mixin';
-
-section.service {
-    @include initial;
-    @include flexDisplay(space-between);
-    flex-wrap: wrap;
-    margin: 2rem 10rem;
-
-    div.serviceContainer {
-        width: 19.5rem;
-        @include border(0.0625rem, rgb(188, 187, 187));
-        border-radius: 1rem;
-        padding: 2.2rem 1.4rem;
-        text-align: center;
-        justify-content: center;
-
-        img.serviceImg {
-            height: 7rem;
-            margin: auto;
-            padding-bottom: 0.9rem;
-        }
-
-        p.serviceDescription {
-            font-size: 0.85rem;
-            color: gray;
-        }
-    }
-}
-
-div.about {
-    margin: 2rem 10rem;
-    text-align: left;
-    font-size: 0.9rem;
-    color: gray;
-    flex-wrap: wrap;
-}
-
-hr {
-    margin: 0 10rem;
-}
