--- src/pages/components/homePage/organisms/header/molecules/cartBtn.scss
@@ -1,37 +1,0 @@
-@import '../../../../../mixin';
-
-button.cartBtn {
-    @include initial;
-    @include flexDisplay(space-evenly);
-    margin-right: 0.9rem;
-    background-color: green;
-    color: white;
-    width: 7.2rem;
-    height: 3rem;
-    border: 0;
-    border-radius: $borderValue*10;
-    cursor: pointer;
-
-    &:hover {
-        background-color: rgb(54, 87, 4);
-    }
-
-    img#cartImg {
-        width: 1.8rem;
-        color: white;
-        -webkit-transform: scaleX(-1);
-        transform: scaleX(-1);
-    }
-
-    div.cartBtnText{
-        font-size: 1rem;
-        font-weight: bold;
-        width: 4.1rem;
-        margin: 0;
-    }
-
-    p {
-        margin: 0;
-    }
-}
-
