--- src/pages/components/homePage/organisms/mainContent/mainProductSection.scss
@@ -1,16 +1,0 @@
-@import '../../../../mixin';
-
-div.content{
-    @include initial;
-    @include flexDisplay(center, normal);
-    aside.cat {
-        @include border();
-        @include initial;
-        width: 16rem;
-    }
-    div.product {
-        display: block;
-        width: 60rem;
-        float: left;
-    }
-}
