+++ src/pages/components/organism/header/molecules/location.scss
@@ -1,0 +1,10 @@
+@import '../../../../mixin';
+
+h1.loacationHeading {
+    font-size: 1.2rem;
+    margin: 0.3rem 0;
+}
+select.locationInput {
+    font-size: 0.9rem;
+    border: 0;
+}
