+++ src/pages/components/organism/header/navbar.scss
@@ -1,0 +1,36 @@
+@import '../../../mixin';
+
+header.navBar {
+    @include initial;
+    @include flexDisplay(space-between);
+    height: fit-content;
+    border-bottom: $borderSide;
+    div {
+        text-align: center;
+        margin: 0 0.5rem;
+    }
+    img#logo {
+        width: 6.5rem;
+        border-right: $borderSide;
+        padding: 0 1.3rem;
+    }
+    input#searchBar {
+        width: 45rem;
+        font-size: 1.2rem;
+        @include border(0.0625rem, gray);
+        border-radius: $borderValue*10;
+        padding: 0.3rem 0.6rem;
+    }
+    a#loginLink {
+        color: rgb(122, 121, 121);
+        font-size: 1.2rem;
+        text-decoration: none;
+        margin: 0 1rem;
+        cursor: pointer;
+
+        &:hover {
+            color: rgb(63, 63, 63);
+        }
+    }
+    
+}
